emulators/winedrop/patches/requirements.txt
pefile
construct==2.5.2
pdbparse
requests
pyyaml
